- generalized controller to be able to handle PID with nominal offset (SO much snappier with nominal offset!) - gathered data and tested PI control with nominal offset - fixed potential future bug in Timers module (not returning the same data time as what was stored (unsigned LONG int now)

+++ Avatar_Micro_Robot/Robot_V2_Accessories/Robotic_Arm_V2/Link_2/trunk/src/Timers.c
@@ -102,7 +102,7 @@ unsigned char IsTimerExpired(unsigned char timerNumber) {
 }
 
 
-unsigned int GetTime(void) {
+unsigned long int GetTime(void) {
 	return currentTime;
 }
